requirements.txt
asgiref==3.7.2
cachetools==5.2.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==2.1.1
cloudinary==1.39.0
decouple==0.0.7
distlib==0.3.8
Django==5.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
inflection==0.5.1
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.8.0
pyparsing==3.0.9
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.1.0
requests==2.28.1
rest-framework-simplejwt==0.0.2
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
timedelta==2020.12.3
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.25.0
